error message on invalid form, moved availabilities and dispositions into separate files
