+++ src/assets/Availabilities.js
@@ -1,0 +1,8 @@
+let availabilities = [
+    'Not Available', 
+    'Available', 
+    'Pending', 
+    'Adopted'
+]
+
+export default availabilities
