+++ src/assets/Dispositions.js
@@ -1,0 +1,7 @@
+const dispositions = [
+    'Good with other animals', 
+    'Good with children', 
+    'Animal must be leased at all times'
+]
+
+export default dispositions
