+++ src/stylesheet.css
@@ -115,7 +115,7 @@ legend {
     margin-bottom: 1rem;
 }
 
-.loginError, .signupError  {
+.loginError, .signupError, .animalError  {
     color: red;
 }
